fibonacci-3

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -12,7 +12,7 @@
 int main(void)
 {
 	int i = 0;
-	long j = 1, k = 2, l = j + k;
+	long double j = 1, k = 2, l = j + k;
 
 	while (i < 98)
 	{
